pyproject.toml
[tool.poetry]
name = "coriolis"
version = "0.1.0"
description = "Small event ticketing and management tool"
license = "AGPL-3.0-or-later"
authors = ["Ryszard Knop <rk@dragonic.eu>"]
repository = "https://github.com/DragoonAethis/Coriolis"
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Beta",
    "Programming Language :: Python :: 3",
    "Environment :: Web Environment",
    "Framework :: Django",
]

[tool.poetry.dependencies]
python = "^3.10"
black = "^23.9.1"
sentry-sdk = "^1.29.2"
Django = "^4.1.9"
django-allauth = "^0.54.0"
django-environ = "^0.10.0"
django-debug-toolbar = "^4.2.0"
django-phonenumber-field = { extras = ["phonenumberslite"], version = "^7.1.0" }
django-crispy-forms = "^2.0"
crispy-bootstrap5 = "^0.7"
gunicorn = "^21.2.0"
whitenoise = { extras = ["brotli"], version = "^6.5.0" }
requests = "^2.31.0"
psycopg2 = "^2.9.7"
django-allauth-2fa = "^0.11.1"
django-colorfield = "^0.9.0"
Pillow = "^9.3.0"
Markdown = "^3.3.6"
django-payments = "^2.0.0"
django-money = "^3.1.0"
dramatiq = { extras = ["watch", "redis"], version = "^1.14.2" }
django-dramatiq = "^0.11.5"
django-hijack = "^3.4.1"
pydantic = "^1.10.12"
xlsxwriter = "^3.1.2"
django-ipware = "^5.0.0"

# Custom fork for Django 4.1 support:
django-dramatiq-email = { git = "https://github.com/DragoonAethis/django-dramatiq-email" }

# Commented out until fixes are upstreamed:
#django-payments-przelewy24 = "0.1.7"


[tool.black]
line-length = 120
extend-exclude = "/events/migrations/*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
